Migrate to list_display for admin rather than str.
