+++ server/auvsi_suas/models/gps_position_test.py
@@ -7,12 +7,6 @@ from django.test import TestCase
 class TestGpsPositionModel(TestCase):
     """Tests the GpsPosition model."""
 
-    def test_str(self):
-        """Tests the str method executes."""
-        pos = GpsPosition(latitude=10, longitude=100)
-        pos.save()
-        self.assertTrue(str(pos))
-
     def assert_distance_equal(self, pos1, pos2, dist, threshold=10):
         """GpsPosition distances are within threshold (ft)."""
         self.assertAlmostEqual(pos1.distance_to(pos2), dist, delta=threshold)
