+++ server/auvsi_suas/models/time_period.py
@@ -45,10 +45,6 @@ class TimePeriod(object):
         self.start = start
         self.end = end
 
-    def __str__(self):
-        """Descriptive text for use in displays."""
-        return 'TimePeriod (start=%s, end=%s)' % (self.start, self.end)
-
     def __eq__(self, other):
         """Two TimePeriods are equal if their attributes are equal."""
         if type(other) == type(self):
